fix(db): dbsetUp bugs and script to create db is added
